Split the exl2 weight ASAP.
