+++ aphrodite/modeling/layers/linear.py
@@ -26,6 +26,68 @@ def adjust_marlin_shard(param, shard_size, shard_offset):
     return shard_size * marlin_tile_size, shard_offset * marlin_tile_size
 
 
+# Split the exl2 weight at group boundary
+def post_init_exl2(layer):
+    q_groups = layer.q_groups
+    num_groups = q_groups.shape[0] // 2
+    input_size = layer.q_invperm.shape[0]
+    tp_rank = get_tensor_model_parallel_rank()
+    tp_size = get_tensor_model_parallel_world_size()
+    rows = 0
+    # row_number, qrow_number, group_number
+    splits = [(0, 0, 0)]
+    index = 1
+    for i in range(num_groups - 1):
+        bits = q_groups[i * 2].item()
+        qrows = (q_groups[i * 2 + 3] - q_groups[i * 2 + 1]).item()
+        rows += qrows * 32 // bits
+
+        if rows >= input_size // tp_size * index:
+            splits.append((rows, q_groups[i * 2 + 3].item(), i + 1))
+            index += 1
+    splits.append((input_size, layer.q_weight.shape[0], num_groups))
+
+    shard_qweight = torch.nn.Parameter(
+        layer.q_weight[splits[tp_rank][1]:splits[tp_rank + 1][1]].clone(),
+        requires_grad=False,
+    )
+    # the outer model.named_parameters() during loading still keeps a reference
+    # so we have to force release memory
+    layer.q_weight.data = torch.empty(0)
+    del layer.q_weight
+    layer.linear_weights["q_weight"] = shard_qweight
+
+    shard_qgroups = torch.nn.Parameter(
+        layer.q_groups[splits[tp_rank][2] * 2:splits[tp_rank + 1][2] *
+                       2].clone(),
+        requires_grad=False,
+    )
+    shard_qgroups[1::2] -= int(shard_qgroups[1])
+    layer.q_groups.data = torch.empty(0)
+    del layer.q_groups
+    layer.linear_weights["q_groups"] = shard_qgroups
+
+    shard_qscale = torch.nn.Parameter(
+        layer.q_scale[splits[tp_rank][2]:splits[tp_rank + 1][2]].clone(),
+        requires_grad=False,
+    )
+    layer.q_scale.data = torch.empty(0)
+    del layer.q_scale
+    layer.linear_weights["q_scale"] = shard_qscale
+
+    shard_qscale_max = torch.nn.Parameter(
+        layer.q_scale_max[splits[tp_rank][2]:splits[tp_rank + 1][2]].clone(),
+        requires_grad=False,
+    )
+    layer.q_scale_max.data = torch.empty(0)
+    del layer.q_scale_max
+    layer.linear_weights["q_scale_max"] = shard_qscale_max
+
+    q_perm = torch.argsort(layer.q_invperm).to(torch.short)
+    layer.linear_weights["q_perm"] = q_perm[splits[tp_rank][0]:splits[tp_rank +
+                                                                      1][0]]
+
+
 class LinearMethodBase(ABC):
     """Base class for different (maybe quantized) linear methods."""
 
@@ -643,6 +705,16 @@ class RowParallelLinear(torch.nn.Module):
         else:
             self.register_parameter("bias", None)
 
+        if self.is_exl2():
+            self.uninitialized = [
+                "q_invperm", "q_weight", "q_scale", "q_scale_max", "q_groups"
+            ]
+
+    def is_exl2(self) -> bool:
+        return hasattr(
+            self.linear_method, "quant_config"
+        ) and self.linear_method.quant_config.get_name() == "exl2"
+
     def weight_loader(self,
                       param: Parameter,
                       loaded_weight: torch.Tensor,
@@ -672,10 +744,18 @@ class RowParallelLinear(torch.nn.Module):
         assert param_data.shape == loaded_weight.shape
         param_data.copy_(loaded_weight)
 
+        if self.is_exl2():
+            weights = self.linear_weights
+            self.uninitialized = [
+                key for key in self.uninitialized if isinstance(
+                    weights[key], torch.nn.parameter.UninitializedParameter)
+            ]
+            if not self.uninitialized:
+                # shard and release memory ASAP to reduce peak memory usage
+                post_init_exl2(self)
+
     def forward(self, input_):
-        is_exl2 = hasattr(
-            self.linear_method, "quant_config"
-        ) and self.linear_method.quant_config.get_name() == "exl2"
+        is_exl2 = self.is_exl2()
         if self.input_is_parallel and is_exl2:
             input_parallel = tensor_model_parallel_all_gather(input_)
         elif self.input_is_parallel or is_exl2:
